feat(identify): include pushed `Info` in `Event::Pushed`
Starting in https://github.com/libp2p/rust-libp2p/pull/3980, we make more use of the identify push protocol by actively notifying the peer when our locally supported protocols change. We emit an event (`Pushed`) to the user in that case. This event however does not include **what** we pushed.

Users might be interested in the internal changes that we push to remote peers. This patch adds the identify `Info` to the `Pushed` variant.

Fixes #4332.

Pull-Request: #4527.