feat(kad): add `refresh_interval` config used to poll `bootstrap`
Previously, users were responsible for calling `bootstrap` on an interval. This was documented but hard to discover for people new to the library. To maintain healthy routing tables, it is advised to regularly call `bootstrap`. By default, we will now do this automatically every 5 minutes and once we add a peer to our routing table, assuming we didn't bootstrap yet. This is especially useful as part of nodes starting up and connecting to bootstrap nodes.

Closes: #4730.

Pull-Request: #4838.

Co-authored-by: stormshield-frb <francois.ribeau.external@stormshield.eu>