feat(kad): expose a kad query facility allowing dynamic num_results (#5555)
## Description

This PR is to expose a kad query facility that allowing specify
num_results dynamically.
It is related to the [Sybil Defence
issue](https://github.com/libp2p/rust-libp2p/issues/4769),
that during the attempt of implementation on higher level code, it is
find will be useful if libp2p-kad can expose such facility.

The PR try not to cause any interference to the existing work flow, only
introduce an `extra exposal`.

## Change checklist

<!-- Please add a Changelog entry in the appropriate crates and bump the
crate versions if needed. See
<https://github.com/libp2p/rust-libp2p/blob/master/docs/release.md#development-between-releases>-->

- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] A changelog entry has been made in the appropriate crates

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>