chore(ci): update mergify script
missed this, with this dependabot is not able to merge PR's by itself as it needs the `send-it` label.
also add a clause for the dismiss of stale approvals to not do so when the last commit was done by one of the `rust-libp2p` maintainers, so that we don't dismiss approvals if the last commit was a merge of master into the PR that is necessary before merging the PR.
Latest example of this happening [here](https://github.com/libp2p/rust-libp2p/pull/5758#issuecomment-2585904938)

Pull-Request: #5809.