feat(relay): emit event when client connections are dropped
When a relay server has no more connection with a reserved client, it would remove the reservation and the drop the circuits without any information passed to the server. It will be useful to for a server to track all its reservations and to know when they're removed (without keeping track of the connections themselves).

This PR aims to notify the server when a reservation closes, with the emission of the following event
```
/// A reservation has been closed.
ReservationClosed { src_peer_id: PeerId },
```

Pull-Request: #5869.