fix(upnp): skip port mapping when an active port mapping is present
This commit aims to fix a bug where a valid listener would be thrown away without attempting a port map. This happens when we have multiple local adapters and the local adapter with the upnp support is not the first one obtained via `NewListenAddr`.
With the new change, we would only skip the listen addr if an `Active Mapping` exists for that same

Pull-Request: #6127.